feat: refactor leave management UI with ShadCN Sheet and improved admin experience
- Remove "New" suffix from component names for consistency
- Implement ShadCN Sheet for employee balance management with proper accessibility
- Add comprehensive leave system configuration in admin settings
- Improve UI design adherence to ShadCN design system standards
- Add proper footer positioning with mt-auto in Sheet layout
- Enhance admin panel with tabs for requests and employee balances
- Update imports and exports to reflect renamed components
- Add proper French labor law compliance indicators and warnings
