+++ src/features/leaves/index.ts
@@ -2,9 +2,7 @@ export { useLeaves } from './hooks/use-leaves'
 export { useLeavesAdmin } from './hooks/use-leaves-admin'
 export { useFrenchLeaveCalculator } from './hooks/use-french-leave-calculator'
 export { LeavesManagement } from './components/LeavesManagement'
-export { LeavesManagementNew } from './components/LeavesManagementNew'
 export { LeavesAdminPanel } from './components/LeavesAdminPanel'
-export { LeavesAdminPanelNew } from './components/LeavesAdminPanelNew'
 export { EmployeeWorkingHoursSettings } from './components/EmployeeWorkingHoursSettings'
 export * from './utils/french-leave-calculator'
 export * from './utils/annual-reset'
